Refactored for less duplication and fixes to docstring and Donation model
